New GdkColorContext object, ported from the XColorContext in XmHTML. It compiles and links, but is *completely* untested.  Feel free to pound on it.
The idea is to do all color management (allocation, etc.) via a
GdkColorContext so that apps will be friendly to 8-bit displays.

GdkColorContext is supposed to work on all visual/depth combinations.
This support, however, is lacking from the rest of Gdk/Gtk.  I will
try to work on that.

 - Federico


git-svn-id: svn+ssh://svn.gnome.org/svn/gtk+/trunk@67 7eb1c76a-c725-0410-a3b5-a37faf6256f8