Iain sent me a xpm that didn't load.  It used "g" instead of "c" in it's color table.
I thought that was borked until I saw Imlib, xv, and Gimp load it up w/o
problem...

So, fixed that (less than one line fix).  HOWEVER, if this is correct,
then the gdk XPM loader in gtk+ also has the same problem.

Owen, got an opinion on this?

Mark


git-svn-id: svn+ssh://svn.gnome.org/svn/gtk+/trunk@2786 7eb1c76a-c725-0410-a3b5-a37faf6256f8